validation_service/requirements.txt
azure-functions==1.21.3
azure-storage-blob==12.19.0
azure-ai-formrecognizer==3.3.2
azure-core==1.38.0
openai==1.12.0
